专题

+++ src/main/java/com/xian/blog/model/Blog.java
@@ -21,7 +21,7 @@ public class Blog {
 	private String summary;// 摘要,内容的前n个字符
 
 	private BlogType blogType; // 博客类型
-	private int typeId;
+	private Long typeId;
 	private String contentNoTag;// 内容无标签,方便分词
 
 	private String thumb;//缩略图
@@ -130,11 +130,11 @@ public class Blog {
 		this.thumb = thumb;
 	}
 
-	public int getTypeId() {
+	public Long getTypeId() {
 		return typeId;
 	}
 
-	public void setTypeId(int typeId) {
+	public void setTypeId(Long typeId) {
 		this.typeId = typeId;
 	}
